Revert "Work on generic IMU driver."
This reverts commit 866ca719c56e5d9c4f61a25e29a3f0375bdb5582.
