+++ Autopilot/AttitudeManager/VN100.c
@@ -17,6 +17,7 @@
 /* Includes ------------------------------------------------------------------*/
 #include "VN100.h"
 #include "VN_lib.h"
+#include "main.h"
 
 
 #ifdef _VN100
