+++ Autopilot/AttitudeManager/VN100.h
@@ -22,7 +22,8 @@
 #include "VN_type.h"
 #include "VN_lib.h"
 
-#define IMU_UPDATE_FREQUENCY 2 //Time in milliseconds
+#define IMU_UPDATE_FREQUENCY 10 //Time in milliseconds
+#define IMU_SPI_PORT 2
 
 /* Exported constants --------------------------------------------------------*/
