+++ Autopilot/AttitudeManager/VN_user.c
@@ -20,6 +20,8 @@
 #include "VN_user.h"
 #include "VN_lib.h"
 #include "delay.h"
+#include "StateMachine.h"
+#include "ProgramStatus.h"
 
 /* Private typedef -----------------------------------------------------------*/
 /* Private define ------------------------------------------------------------*/
