+++ Autopilot/Common/Interfaces/SPI.c
@@ -204,12 +204,12 @@ void initSPI(uint8_t interface, uint16_t clock, spi_mode mode, spi_width width, 
 byte SPI_TX_RX(uint8_t interface, byte data) {
     if (interface == 1) {
         SPI1BUF = data;
-        asm volatile("nop");
+        while (SPI1STATbits.SPITBF){}
         while (!SPI1STATbits.SPIRBF){}
         return SPI1BUF;
     } else if (interface == 2) {
         SPI2BUF = data;
-        asm volatile("nop");
+        while (SPI2STATbits.SPITBF){}
         while (!SPI2STATbits.SPIRBF){}
         return SPI2BUF;
     }
